requirements.txt
pynput>=1.7.7
pyaudio>=0.2.14
faster-whisper>=1.0.0
opencc-python-reimplemented
pyperclip>=1.9.0
rumps
pyinstaller